refactor(ir): move random() to api.py
