+++ ibis/expr/types/category.py
@@ -1,11 +1,33 @@
+from typing import Optional, Sequence
+
 from public import public
 
+import ibis.expr.operations as ops
 from ibis.expr.types.generic import Column, Scalar, Value
+from ibis.expr.types.strings import StringValue
 
 
 @public
 class CategoryValue(Value):
-    pass  # noqa: E701,E302
+    def label(
+        self, labels: Sequence[str], nulls: Optional[str] = None
+    ) -> StringValue:
+        """Format a known number of categories as strings.
+
+        Parameters
+        ----------
+        labels
+            Labels to use for formatting categories
+        nulls
+            How to label any null values among the categories
+
+        Returns
+        -------
+        StringValue
+            Labeled categories
+        """
+        op = ops.CategoryLabel(self, labels, nulls)
+        return op.to_expr()
 
 
 @public
